Refactor to consolidate api methods
